requirements.txt
flask==2.3.3
boto3==1.28.42
web3==6.9.0
cryptography==41.0.3
PyYAML==6.0.1